Track employment rate in records
