+++ src/main.rs
@@ -2,11 +2,11 @@
 #![allow(unused_imports)]
 #![allow(unused_variables)]
 
-use bevy_app::prelude::*;
-use bevy_ecs::prelude::*;
-use bevy_time::{Time, Real};
 #[cfg(feature = "graphics")]
 use bevy::prelude::DefaultPlugins;
+use bevy_app::prelude::*;
+use bevy_ecs::prelude::*;
+use bevy_time::{Real, Time};
 
 mod baby_spawner;
 mod gregslist;
@@ -22,7 +22,7 @@ mod view;
 use crate::baby_spawner::{BabySpawnerConfig, BabySpawnerPlugin};
 use crate::inventory::InventoryPlugin;
 use crate::mortality::system::apply_mortality_with_rate;
-use crate::records::{Records, rolling_mean::RollingMean};
+use crate::records::{rolling_mean::RollingMean, Records};
 use jobs::Job;
 
 const SEC: f64 = 1.0;
@@ -86,6 +86,7 @@ fn main() {
             deaths: 0,
             birth_rate: RollingMean::new(DAY),
             death_rate: RollingMean::new(DAY),
+            employment_rate: 0.0,
         })
         .add_systems(Update, {
             let deaths_per_sec_per_person = SPEED / (AVERAGE_LIFESPAN_YEARS * YR);
