+++ src/records/mod.rs
@@ -4,8 +4,10 @@ pub mod rolling_mean;
 #[cfg(feature = "graphics")]
 pub mod ui;
 
-pub use self::records::{record_births, record_deaths, Records};
+pub use self::records::{record_births, record_deaths, record_employment_rate, Records};
 pub use self::rolling_mean::RollingMean;
 #[cfg(feature = "graphics")]
-pub use self::ui::{spawn_population_text, update_population_text};
+pub use self::ui::{
+    spawn_employment_text, spawn_population_text, update_employment_text, update_population_text,
+};
 pub use plugin::RecordsPlugin;
