+++ src/records/plugin.rs
@@ -1,9 +1,9 @@
-use crate::records::Records;
 #[cfg(feature = "graphics")]
 use crate::records::ui::{
     spawn_employment_text, spawn_population_text, update_employment_text, update_population_text,
 };
-use crate::records::{record_births, record_deaths};
+use crate::records::Records;
+use crate::records::{record_births, record_deaths, record_employment_rate};
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
 use bevy_time::prelude::*;
@@ -16,6 +16,9 @@ impl Plugin for RecordsPlugin {
         app.add_systems(Startup, (spawn_population_text, spawn_employment_text))
             .add_systems(Update, (update_population_text, update_employment_text));
 
-        app.add_systems(Update, (record_births, record_deaths));
+        app.add_systems(
+            Update,
+            (record_births, record_deaths, record_employment_rate),
+        );
     }
 }
