+++ src/records/records.rs
@@ -1,9 +1,11 @@
 use crate::baby_spawner::BabyBorn;
+use crate::hiring_manager::component::Unemployed;
 use crate::mortality::Death;
+use crate::person::Person;
 use crate::records::RollingMean;
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
-use bevy_time::{Time, Real};
+use bevy_time::{Real, Time};
 
 #[derive(Resource, Debug, Clone)]
 pub struct Records {
@@ -11,6 +13,7 @@ pub struct Records {
     pub deaths: usize,
     pub birth_rate: RollingMean,
     pub death_rate: RollingMean,
+    pub employment_rate: f32,
 }
 
 impl Records {
@@ -40,5 +43,20 @@ pub fn record_deaths(
     for _ in deaths.read() {
         records.deaths = records.deaths.saturating_add(1);
         records.death_rate.push(now);
+    }
+}
+
+pub fn record_employment_rate(
+    mut records: ResMut<Records>,
+    people: Query<Entity, With<Person>>,
+    unemployed: Query<Entity, With<Unemployed>>,
+) {
+    let total = people.iter().count();
+    let jobless = unemployed.iter().count();
+    if total > 0 {
+        let employed = total.saturating_sub(jobless);
+        records.employment_rate = employed as f32 / total as f32;
+    } else {
+        records.employment_rate = 0.0;
     }
 }
